refactor layout
