+++ src/hardcopy_pango.h
@@ -1,4 +1,19 @@
-/* cairo.c */
+/* vi:set ts=8 sts=4 sw=4 noet:
+ *
+ * VIM - Vi IMproved	by Bram Moolenaar
+ *
+ * Do ":help uganda"  in Vim to read copying and usage conditions.
+ * Do ":help credits" in Vim to see a list of people who contributed.
+ * See README.txt for an overview of the Vim source code.
+ */
+
+#ifndef HARDCOPY_PANGO_H
+#define HARDCOPY_PANGO_H
+
+#include "vim.h"
+
+#if defined(FEAT_PRINT_PANGO)
+
 int mch_print_init(prt_settings_T *psettings, char_u *jobname, int forceit);
 int mch_print_begin(prt_settings_T *psettings);
 void mch_print_end(prt_settings_T *psettings);
@@ -7,8 +22,11 @@ int mch_print_begin_page(char_u *str);
 int mch_print_blank_page(void);
 void mch_print_start_line(int margin, int page_line);
 int mch_print_text_out(char_u *textp, int len);
-void mch_print_set_font(int iBold, int iItalic, int iUnderline);
+void mch_print_set_font(int bold, int italic, int underline);
 void mch_print_set_bg(long_u bgcol);
 void mch_print_set_fg(long_u fgcol);
 void mch_print_cleanup(void);
-/* vim: set ft=c : */
+
+#endif
+
+#endif
